Refactor math in #chance method
While this is a pretty simple refactor to make use of taking in dice
rolls as an array and to make use of an enumerable, it also belies the
beginning of decision-making regarding how to handle the mixed use of
instance and class methods for methods which are only superficially
different.

While the spec suggests only "minimal" changes to the public interface,
it also suggests that the interface be consistent and simple, which it
currently is not. Hence, future commits will reflect decision-making
regarding whether to keep instance or class methods, and will be single
commits for each method changed.