Refactor #ones method to use enumerable
We can already see that there will be a lot of repetition for the
methods calculating the number of times a specific number appears in a
series of rolls. The pattern in this commit will form the basis for
refactoring out a common method to be used by all the numbers.